rename to LineTraceParamPass.
